feat(settings): move user recipes management to user settings
