+++ src/routes/recipes/mod.rs
@@ -1,4 +1,3 @@
-pub mod community;
 pub mod detail;
 pub mod edit;
 pub mod import;
