+++ src/routes/settings/mod.rs
@@ -1,3 +1,4 @@
 pub mod account;
 pub mod billing;
 pub mod general;
+pub mod recipes;
