+++ templates/_recipes.html
@@ -40,7 +40,7 @@
 
                     <!-- Quick Actions -->
                     <div class="mt-6 pt-6 border-t space-y-2">
-                        <button ts-req="/recipes/create" ts-req-method="post" class="block w-full px-4 py-2 bg-green-600 text-white font-semibold rounded-lg hover:bg-green-700 text-center text-sm">
+                        <button ts-req="/settings/recipes/create" ts-req-method="post" class="block w-full px-4 py-2 bg-green-600 text-white font-semibold rounded-lg hover:bg-green-700 text-center text-sm">
                             ✏️ {{ "Create Recipe"|t }}
                         </button>
                         <a href="/recipes/import" class="block w-full px-4 py-2 bg-blue-600 text-white font-semibold rounded-lg hover:bg-blue-700 text-center text-sm">
